Implemented ACTUAL data logging
Basically the issue with the previous idea was that it would only update the local files and not the actual github files, meaning the data would die on reboot. For this to work we need to periodically pull the data logs off of the RIO, or just have a usb stick formatted to FAT32 in.